+++ wannierberri/__init__.py
@@ -35,7 +35,7 @@ from .parallel import Parallel, Serial
 from .smoother import get_smoother
 from .evaluate_k import evaluate_k
 from . import utils
-from . import data_K
+# from . import data_K
 from .result.tabresult import npz_to_fermisurfer
 from termcolor import cprint
